Home changes-- Spacemacs and copier

+++ copier.sh
@@ -1,18 +1,14 @@
 #!/bin/zsh
-############################
-# .make.sh
-# This script creates symlinks from the home directory to any desired dotfiles in ~/dotfiles
-############################
 
-########## Variables
-
-dir=~/.dotfiles/                    # dotfiles directory
+dir=~/dotfiles/                    # dotfiles directory
 
 cd $dir
 
 for file in $dir/.*; do
     base=$(basename $file)
-    echo $base
-    rm ~/$base
-    ln -s $file ~/$basename
+    if [[ "$base" != ".git" ]]; then
+           echo $base
+           rm ~/$base
+           ln -s $file ~/$basename
+    fi
 done
